Move 'app.js' to 'src' directory
