+++ package.json
@@ -4,10 +4,10 @@
   "description": "Coding challenge in Node.js",
   "license": "WTFPL",
   "author": "",
-  "main": "app.js",
+  "main": "src/app",
   "repository": "https://github.com/busbud/coding-challenge-backend-c.git",
   "scripts": {
-    "start": "babel-node app",
+    "start": "babel-node src/app",
     "test": "standard-busbud && mocha --compilers js:babel/register"
   },
   "dependencies": {
